use floats for audio buffers

+++ engine/audio_feed.h
@@ -7,7 +7,7 @@
 
 typedef struct {
   PaUtilRingBuffer* ringbuffer;
-  double* ringbuffer_data;
+  float* ringbuffer_data;
 } AudioFeed;
 
 
@@ -16,7 +16,7 @@ AudioFeed* AudioFeed_create();
 void AudioFeed_destroy();
 
 // Read the newest data from the ring buffer
-void AudioFeed_read(AudioFeed* af, double* output, int output_length);
+void AudioFeed_read(AudioFeed* af, float* output, int output_length);
 
 // write into the ring buffer
-void AudioFeed_process(AudioFeed* af, double* input, int input_length);
+void AudioFeed_process(AudioFeed* af, float* input, int input_length);
